# Conflicts:
#	arduino/wheeledbase/Makefile
#	arduino/wheeledbase/instructions.h
#	arduino/wheeledbase/wheeledbase.ino
#	esp32/wheeledbase/Makefile
#	esp32/wheeledbase/instructions.h
#	esp32/wheeledbase/wheeledbase.ino
#	raspberrypi/daughter_cards/wheeledbase.py

+++ esp32/common/IK/Matrix.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 
 
-typedef double** matrix_t;
+typedef float** matrix_t;
 
 namespace IK
 {
@@ -12,13 +12,13 @@ namespace IK
 class Matrix3
 {	
 	public:
-		matrix_t createMatrix33(double X1, double X2, double X3, double Y1, double Y2, double Y3, double Z1, double Z2, double Z3);
-		matrix_t createMatrix31(double X1, double Y1, double Z1);
+		matrix_t createMatrix33(float X1, float X2, float X3, float Y1, float Y2, float Y3, float Z1, float Z2, float Z3);
+		matrix_t createMatrix31(float X1, float Y1, float Z1);
 
 		matrix_t multMatrix33x13(matrix_t mat33, matrix_t mat13);
 
-		double norm(matrix_t mat);
-		double det(matrix_t mat);
+		float norm(matrix_t mat);
+		float det(matrix_t mat);
 		matrix_t solve(matrix_t mat33, matrix_t mat13);
 
 		void free(matrix_t m);
@@ -27,13 +27,13 @@ class Matrix3
 class Matrix2
 {
 	public:
-		matrix_t createMatrix22(double X1, double X2, double Y1, double Y2);
-		matrix_t createMatrix21(double X1, double Y1);
+		matrix_t createMatrix22(float X1, float X2, float Y1, float Y2);
+		matrix_t createMatrix21(float X1, float Y1);
 
 		matrix_t multMatrix22x12(matrix_t mat22, matrix_t mat12);
 
-		double norm(matrix_t mat);
-		double det(matrix_t mat);
+		float norm(matrix_t mat);
+		float det(matrix_t mat);
 		matrix_t solve(matrix_t mat22, matrix_t mat12);
 		
 		void free(matrix_t m);
